requirements.txt
tavily-python
openai
streamlit
python-dotenv
